+++ src/snowflake/snowpark/modin/plugin/docstrings/series_utils.py
@@ -1525,11 +1525,67 @@ class CombinedDatetimelikeProperties:
 
     @property
     def is_quarter_start():
-        pass
+        """
+        Indicator for whether the date is the first day of a quarter.
+
+        Returns
+        -------
+        is_quarter_start : Series or DatetimeIndex
+            The same type as the original data with boolean values. Series will have the same name and index. DatetimeIndex will have the same name.
+
+        See also
+        --------
+        quarter
+            Return the quarter of the date.
+        is_quarter_end
+            Similar property for indicating the quarter end.
+
+        Examples
+        --------
+        This method is available on Series with datetime values under the .dt accessor, and directly on DatetimeIndex.
+
+        >>> df = pd.DataFrame({'dates': pd.date_range("2017-03-30",
+        ...                   periods=4)})
+        >>> df.assign(quarter=df.dates.dt.quarter,
+        ...           is_quarter_start=df.dates.dt.is_quarter_start)
+            dates  quarter  is_quarter_start
+        0 2017-03-30        1             False
+        1 2017-03-31        1             False
+        2 2017-04-01        2              True
+        3 2017-04-02        2             False
+        """
 
     @property
     def is_quarter_end():
-        pass
+        """
+        Indicator for whether the date is the last day of a quarter.
+
+        Returns
+        -------
+        is_quarter_end : Series or DatetimeIndex
+            The same type as the original data with boolean values. Series will have the same name and index. DatetimeIndex will have the same name.
+
+        See also
+        --------
+        quarter
+            Return the quarter of the date.
+        is_quarter_start
+            Similar property indicating the quarter start.
+
+        Examples
+        --------
+        This method is available on Series with datetime values under the .dt accessor, and directly on DatetimeIndex.
+
+        >>> df = pd.DataFrame({'dates': pd.date_range("2017-03-30",
+        ...                    periods=4)})
+        >>> df.assign(quarter=df.dates.dt.quarter,
+        ...           is_quarter_end=df.dates.dt.is_quarter_end)
+            dates  quarter    is_quarter_end
+        0 2017-03-30        1             False
+        1 2017-03-31        1              True
+        2 2017-04-01        2             False
+        3 2017-04-02        2             False
+        """
 
     @property
     def is_year_start():