+++ tests/unit/modin/test_series_dt.py
@@ -35,8 +35,6 @@ def mock_query_compiler_for_dt_series() -> SnowflakeQueryCompiler:
         (lambda s: s.dt.time, "time"),
         (lambda s: s.dt.timetz, "timetz"),
         (lambda s: s.dt.weekday, "weekday"),
-        (lambda s: s.dt.is_quarter_start, "is_quarter_start"),
-        (lambda s: s.dt.is_quarter_end, "is_quarter_end"),
         (lambda s: s.dt.is_leap_year, "is_leap_year"),
         (lambda s: s.dt.daysinmonth, "daysinmonth"),
         (lambda s: s.dt.days_in_month, "days_in_month"),